refactor: extract duplication suppression helpers
